+++ app/src/js/paths/groups.js
@@ -6,7 +6,7 @@ const groupRoutes = [
 ];
 
 const singleGroupRoutes = [
-  [strings.back_to_submissions, '/requests', 'sidebar__nav--back'],
+  [strings.back_to_requests, '/requests', 'sidebar__nav--back'],
   [strings.back_to_groups, null, 'sidebar__nav--back']
 ];