+++ app/src/js/utils/table-config/requests.js
@@ -6,7 +6,7 @@ import {
   shortDateNoTimeYearFirst,
   nullValue,
   fromNow,
-  submissionLink,
+  requestLink,
   bool
 } from '../format';
 import {
@@ -25,9 +25,15 @@ export const getPrivileges = () => {
       ? privileges.find(o => o.match(/ADMIN/g))
       : roles.find(o => o.short_name.match(/manager/g)),
     isAdmin: privileges.find(o => o.match(/ADMIN/g)),
+    isProducer: privileges.find(o => o.match(/ADMIN/g))
+      ? privileges.find(o => o.match(/ADMIN/g))
+      : roles.find(o => o.short_name.match(/data_producer/g)),
     canReview: privileges.find(o => o.match(/ADMIN/g))
       ? privileges.find(o => o.match(/ADMIN/g))
       : privileges.find(o => o.match(/REQUEST_REVIEW/g)),
+    canReassign: privileges.find(o => o.match(/ADMIN/g))
+      ? privileges.find(o => o.match(/ADMIN/g))
+      : privileges.find(o => o.match(/REQUEST_REASSIGN/g)),
     canCreateForm: privileges.find(o => o.match(/ADMIN/g))
       ? privileges.find(o => o.match(/ADMIN/g))
       : privileges.find(o => o.match(/FORM_CREATE/g)),
@@ -46,10 +52,13 @@ export const newLink = (request, formalName) => {
   } else {
     disabled = true;
   }
-  const disabledClass = disabled ? 'button--disabled' : '';
-  // This element was purposefully left as an anchor tag (rather than react Link) since the page is redirected away from
-  // the dashboard site to the forms site. Converting to a Link component will result in a malformed url.
-  return <a href={request} className={`button button--small button--green form-group__element--left button--no-icon ${disabledClass}`} aria-label={formalName || 'take action'}>{formalName}</a>;
+  if (disabled) {
+    return <Link to={''} className={'button button--medium button--clear form-group__element--left button--no-icon'} aria-label={formalName}>{formalName}</Link>;
+  } else {
+    // This element was purposefully left as an anchor tag (rather than react Link) since the page is redirected away from
+    // the dashboard site to the forms site. Converting to a Link component will result in a malformed url.
+    return <a href={request} className={'button button--medium button--green form-group__element--left button--no-icon'} aria-label={formalName || 'take action'}>{formalName}</a>;
+  }
 };
 
 export const assignWorkflow = (request, formalName) => {
@@ -60,9 +69,12 @@ export const assignWorkflow = (request, formalName) => {
   } else {
     disabled = true;
   }
-  const disabledClass = disabled ? 'button--disabled' : '';
-  return <Link className={`button button--small button--green form-group__element--left button--no-icon ${disabledClass}`}
+  if (disabled) {
+    return <Link to={''} className={'button button--medium button--clear form-group__element--left button--no-icon'} aria-label={formalName}>{formalName}</Link>;
+  } else {
+    return <Link className={'button button--medium button--green form-group__element--left button--no-icon'}
                to={`${request}`} name={'assignButton'} aria-label={formalName || 'assign workflow'}>{formalName}</Link>;
+  }
 };
 
 export const existingLink = (row, formId, formalName, step) => {
@@ -73,11 +85,14 @@ export const existingLink = (row, formId, formalName, step) => {
   } else {
     disabled = true;
   }
-  const disabledClass = disabled ? 'button--disabled' : '';
-  if (typeof formId === 'undefined') {
-    return <Link to={`/requests/approval?requestId=${row.id}&step=${step}`} className={`button button--small button--green form-group__element--left button--no-icon ${disabledClass}`} aria-label={formalName}>{formalName || 'review item'}</Link>;
+  if (disabled) {
+    return <Link to={''} className={'button button--medium button--clear form-group__element--left button--no-icon'} aria-label={formalName}>{formalName}</Link>;
   } else {
-    return <Link to={`/forms/id/${formId}?requestId=${row.id}`} className={`button button--small button--green form-group__element--left button--no-icon ${disabledClass}`} aria-label={formalName || 'view form details'}>{formalName}</Link>;
+    if (typeof formId === 'undefined') {
+      return <Link to={`/requests/approval?requestId=${row.id}&step=${step}`} className={'button button--medium button--green form-group__element--left button--no-icon'} aria-label={formalName || 'review item'}>{formalName}</Link>;
+    } else {
+      return <Link to={`/forms/id/${formId}?requestId=${row.id}`} className={'button button--medium button--green form-group__element--left button--no-icon'} aria-label={formalName || 'view form details'}>{formalName}</Link>;
+    }
   }
 };
 
@@ -219,7 +234,7 @@ export const errorTableColumns = [
   {
     Header: 'Requests',
     accessor: (row) => row.id,
-    Cell: row => row.row ? submissionLink(row.row.original.id) : null,
+    Cell: row => row.row ? requestLink(row.row.original.id) : null,
     id: 'id',
     width: 200
   },